Avoid excessive privilege on set/query file info/security info (#277)
* Avoid excessive privilege on set/query file info/security info

Originally, we are directly using GENERIC_READ/GENERIC_WRITE and this would cause excessive privilege. This commit is avoiding the excessive privilege issue on set or query for both file information class and the security information class.

* Fix : Handle set group owner on security info case

Avoid excessive privilege commit didn't handle the case to set group owner on security info case. This is a fix to handle it back. And also start with a empty access mask set on setSecurityInfo method.